iam auth: avoid ydb/library/yql->kqp dependency
